requirements.txt
asgiref==3.5.2
colorama==0.4.5
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.6
django-cache-url==3.4.2
django-widgets-improved==1.5.0
djangorestframework==3.13.1
environs==8.0.0
marshmallow==3.17.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
qrcode==7.3.1
sqlparse==0.4.2
tzdata==2022.1
